Rename metrics to dataset
