+++ README.md
@@ -15,7 +15,7 @@ pip install -e .
 A small example to get started:
 
 ```python
-from clin_nlp_metrics.metrics import Dataset
+from clin_nlp_metrics.dataset import Dataset
 import json
 
 # medcattrainer
@@ -23,7 +23,7 @@ import json
 
 with open('medcattrainer_export.json', 'rb') as f:
     mtrainer_data = json.load(f)
-    
+
 d1 = Dataset.from_medcattrainer(mctrainer_data)
 
 # clinlp
@@ -31,6 +31,7 @@ import clinlp
 import spacy
 
 from model import get_model  # not included
+
 nlp = get_model()
 nlp_docs = nlp.pipe([doc['text'] for doc in data['projects'][0]['documents']])
