+++ src/clin_nlp_metrics/__init__.py
@@ -1,0 +1,1 @@
+from .dataset import Dataset
